feat: live twitter feed based on selector
